Add user existence check in Diagnostics

+++ database/migrations/2020_12_12_203153_migrate_admin_user.php
@@ -33,7 +33,9 @@ class MigrateAdminUser extends Migration
 	{
 		if (Schema::hasTable('users')) {
 			$user = User::find(0);
-			$user->delete();
+			if ($user != null) {
+				$user->delete();
+			}
 		}
 	}
 }
